Remove /v2 from paths
In practical implementation, a server will provide a base URL indicating the version as appropriate.  Prescribing the form of the paths in the API limits design space of the server implementations and/or makes URLs more awkward (e.g., example.com/rid/v21/v2/dss/subscriptions rather than allowing example.com/rid/v21/dss/subscriptions).  Ideally, the /dss and /uss prefixes would be removed as well, but this would cause path conflicts when the DSS and USS endpoints are defined in the same OpenAPI file.